requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.12.5
langchain==0.1.16
langchain-community==0.0.33
langchain-openai==0.1.3
langchain-google-genai==1.0.3
httpx>=0.28.0,<0.29.0
python-dotenv==1.0.1
redis==5.0.1
daytona
langgraph
